Upload migration script and cleaned up CSV
The previous CSV had valid javascript, but invalid JSON on it as part of the dataset. I wrote a migration script that took the javascript objects from the .JS file and wrote them into sqlite, which could then output the resulting CSV. The CSV added here is the result of such a migration.